tidying up of datatypes and adding notion of exercise week

+++ backend/sql/BIGGYM/code/api/spGetIdForExercise.sql
@@ -15,7 +15,7 @@ drop procedure if exists spGetIdForExercise;
 delimiter $$
 create procedure spGetIdForExercise(in vExerciseName varchar(128),
                                     in vBodyPartName varchar(128),
-                                   out ObjectId smallint,
+                                   out ObjectId mediumint unsigned,
                                    out ReturnCode int)
 begin
